requirements.txt
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
Pillow==7.1.2